requirements.txt
build==0.10.0
certifi==2023.5.7
numpy==1.25.0
packaging==23.1
pygame==2.5.0
pyproject_hooks==1.0.0
python-dotenv==1.0.0
